requirements.txt
py3-bencode==0.0.3
aio-udp-server==0.0.2
chardet==3.0.4
motor==1.1
pymongo==3.5.1
